# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Catalog of indie games — browser or download, free, paid or subscription.",
        "title": "IndieForge API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/admin/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get service settings (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/indieforge_internal_dto.Settings"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update service settings (admin)",
                "parameters": [
                    {
                        "description": "Settings",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indieforge_internal_dto.Settings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/indieforge_internal_dto.Settings"
                            }
                        }
                    }
                }
            }
        },
...
        "/moderation/reports": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderation"
                ],
                "summary": "List reports (moderator+)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "open|resolved|dismissed",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/moderation/reports/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderation"
                ],
                "summary": "Get a report (moderator+)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/indieforge_internal_dto.ReportDTO"
                            }
                        }
                    }
                }
            }
        },
        "/moderation/reports/{id}/resolve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderation"
                ],
                "summary": "Resolve a report (dismiss | hide-game | remove-game)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Action",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indieforge_internal_dto.ResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/indieforge_internal_dto.ReportDTO"
                            }
                        }
                    }
                }
            }
        },
...
        "/reports": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderation"
                ],
                "summary": "Report a game",
                "parameters": [
                    {
                        "description": "Report",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indieforge_internal_dto.CreateReportRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/indieforge_internal_dto.ReportDTO"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "indieforge_internal_dto.CreateReportRequest": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "targetId": {
                    "type": "string"
                },
                "targetType": {
                    "type": "string"
                }
            }
        },
...
        "indieforge_internal_dto.ReportDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "handledBy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "reporterId": {
                    "type": "string"
                },
                "resolution": {
                    "type": "string"
                },
                "resolvedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "targetId": {
                    "type": "string"
                },
                "targetType": {
                    "type": "string"
                }
            }
        },
        "indieforge_internal_dto.ResolveRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                }
            }
        },
        "indieforge_internal_dto.Settings": {
            "type": "object",
            "properties": {
                "commissionPercent": {
                    "type": "integer"
                },
                "popularEnabled": {
                    "type": "boolean"
                },
                "trendingEnabled": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}